# docs/api/openapi-v1.yaml
...
openapi: 3.1.0
info:
  title: PigOS API
  description: 'Global Swine Farm Management SaaS — Base API + AI Addon platform.


    이 파일은 라이브 라우트에서 자동 생성됨 (scripts/gen_openapi.py). 수기 편집 금지.

    사람용 연동 계약서: docs/mobile-integration-contract.md


    인증: Bearer JWT (/api/v1/auth/login). access 15분 / refresh 7일.

    멀티테넌트: 모든 농장 데이터는 farm_id로 격리.'
  version: 0.1.0
  contact:
    name: PigOS Support
    url: https://pigos.io
  license:
    name: Proprietary
paths:
...
  /api/v1/farms/{farm_id}/finishers:
    get:
      tags:
      - Finishers
      summary: List Finisher Groups
      operationId: list_finisher_groups_api_v1_farms__farm_id__finishers_get
      security:
      - HTTPBearer: []
      parameters:
      - name: farm_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          description: Farm UUID
          title: Farm Id
        description: Farm UUID
      - name: active_only
        in: query
        required: false
        schema:
          type: boolean
          description: True = 출하 전 그룹만
          default: false
          title: Active Only
        description: True = 출하 전 그룹만
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FinisherGroupResponse'
                title: Response List Finisher Groups Api V1 Farms  Farm Id  Finishers
                  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Finishers
      summary: Create Finisher Group
      description: 비육돈 그룹 입식 등록
      operationId: create_finisher_group_api_v1_farms__farm_id__finishers_post
      security:
      - HTTPBearer: []
      parameters:
      - name: farm_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          description: Farm UUID
          title: Farm Id
        description: Farm UUID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FinisherGroupCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinisherGroupResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/farms/{farm_id}/finishers/{group_id}/ship:
    post:
      tags:
      - Finishers
      summary: Ship Finisher Group
      description: 출하 처리 — 그룹 종료
      operationId: ship_finisher_group_api_v1_farms__farm_id__finishers__group_id__ship_post
      security:
      - HTTPBearer: []
      parameters:
      - name: group_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Group Id
      - name: farm_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          description: Farm UUID
          title: Farm Id
        description: Farm UUID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FinisherGroupShip'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinisherGroupResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/farms/{farm_id}/finishers/{group_id}:
    delete:
      tags:
      - Finishers
      summary: Delete Finisher Group
      operationId: delete_finisher_group_api_v1_farms__farm_id__finishers__group_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: group_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Group Id
      - name: farm_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          description: Farm UUID
          title: Farm Id
        description: Farm UUID
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - Finishers
      summary: Update Finisher Group
      operationId: update_finisher_group_api_v1_farms__farm_id__finishers__group_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: group_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Group Id
      - name: farm_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          description: Farm UUID
          title: Farm Id
        description: Farm UUID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FinisherGroupUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinisherGroupResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/farms/{farm_id}/piglets:
    get:
      tags:
      - Piglets
      summary: List Piglet Groups
      operationId: list_piglet_groups_api_v1_farms__farm_id__piglets_get
      security:
      - HTTPBearer: []
      parameters:
      - name: farm_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          description: Farm UUID
          title: Farm Id
        description: Farm UUID
      - name: active_only
        in: query
        required: false
        schema:
          type: boolean
          description: True = 전출 전 그룹만
          default: false
          title: Active Only
        description: True = 전출 전 그룹만
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PigletGroupResponse'
                title: Response List Piglet Groups Api V1 Farms  Farm Id  Piglets
                  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Piglets
      summary: Create Piglet Group
      description: 자돈 그룹 시작 (이유 후 그룹 등록) — 피그플랜 P03010100 자돈그룹시작
      operationId: create_piglet_group_api_v1_farms__farm_id__piglets_post
      security:
      - HTTPBearer: []
      parameters:
      - name: farm_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          description: Farm UUID
          title: Farm Id
        description: Farm UUID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PigletGroupCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PigletGroupResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/farms/{farm_id}/piglets/{group_id}/deaths:
    post:
      tags:
      - Piglets
      summary: Record Deaths
      description: 자돈 폐사 누적 기록
      operationId: record_deaths_api_v1_farms__farm_id__piglets__group_id__deaths_post
      security:
      - HTTPBearer: []
      parameters:
      - name: group_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Group Id
      - name: farm_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          description: Farm UUID
          title: Farm Id
        description: Farm UUID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PigletGroupDeathRecord'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PigletGroupResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/farms/{farm_id}/piglets/{group_id}/transfer:
    post:
      tags:
      - Piglets
      summary: Transfer Out Piglet Group
      description: 전출/판매 처리 — 피그플랜 P03010400 자돈전출/출하기록
      operationId: transfer_out_piglet_group_api_v1_farms__farm_id__piglets__group_id__transfer_post
      security:
      - HTTPBearer: []
      parameters:
      - name: group_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Group Id
      - name: farm_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          description: Farm UUID
          title: Farm Id
        description: Farm UUID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PigletGroupTransferOut'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PigletGroupResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/farms/{farm_id}/piglets/transfers:
    post:
      tags:
      - Piglets
      summary: Create Piglet Transfer
      description: '양자/대리모 기록.

        분만 기록과 별도 — PSY는 생물학적 산자 기준, 이유두수는 양자 반영 수 기준.'
      operationId: create_piglet_transfer_api_v1_farms__farm_id__piglets_transfers_post
      security:
      - HTTPBearer: []
      parameters:
      - name: farm_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          description: Farm UUID
          title: Farm Id
        description: Farm UUID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PigletTransferCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PigletTransferResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - Piglets
      summary: List Piglet Transfers
      operationId: list_piglet_transfers_api_v1_farms__farm_id__piglets_transfers_get
      security:
      - HTTPBearer: []
      parameters:
      - name: farm_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          description: Farm UUID
          title: Farm Id
        description: Farm UUID
      - name: sow_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          description: 특정 모돈 필터
          title: Sow Id
        description: 특정 모돈 필터
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PigletTransferResponse'
                title: Response List Piglet Transfers Api V1 Farms  Farm Id  Piglets
                  Transfers Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    FinisherGroupCreate:
      properties:
        group_code:
          type: string
          maxLength: 30
          title: Group Code
          description: '그룹 코드 (예: FG-2026-001)'
        batch_name:
          anyOf:
          - type: string
            maxLength: 100
          - type: 'null'
          title: Batch Name
        start_date:
          type: string
          format: date
          title: Start Date
        head_count_in:
          type: integer
          minimum: 1.0
          title: Head Count In
          description: 입식 두수
        avg_entry_weight_kg:
          anyOf:
          - type: number
            exclusiveMinimum: 0.0
          - type: 'null'
          title: Avg Entry Weight Kg
          description: 평균 입식 체중 (kg)
        building_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Building Id
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
      type: object
      required:
      - group_code
      - start_date
      - head_count_in
      title: FinisherGroupCreate
    FinisherGroupResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        farm_id:
          type: string
          format: uuid
          title: Farm Id
        group_code:
          type: string
          title: Group Code
        batch_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Batch Name
        start_date:
          type: string
          format: date
          title: Start Date
        end_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: End Date
        head_count_in:
          type: integer
          title: Head Count In
        head_count_out:
          anyOf:
          - type: integer
          - type: 'null'
          title: Head Count Out
        avg_entry_weight_kg:
          anyOf:
          - type: number
          - type: 'null'
          title: Avg Entry Weight Kg
        avg_exit_weight_kg:
          anyOf:
          - type: number
          - type: 'null'
          title: Avg Exit Weight Kg
        building_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Building Id
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - farm_id
      - group_code
      - batch_name
      - start_date
      - end_date
      - head_count_in
      - head_count_out
      - avg_entry_weight_kg
      - avg_exit_weight_kg
      - building_id
      - notes
      - created_at
      - updated_at
      title: FinisherGroupResponse
    FinisherGroupShip:
      properties:
        end_date:
          type: string
          format: date
          title: End Date
        head_count_out:
          type: integer
          minimum: 1.0
          title: Head Count Out
          description: 출하 두수
        avg_exit_weight_kg:
          anyOf:
          - type: number
            exclusiveMinimum: 0.0
          - type: 'null'
          title: Avg Exit Weight Kg
          description: 평균 출하 체중 (kg)
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
      type: object
      required:
      - end_date
      - head_count_out
      title: FinisherGroupShip
      description: 출하 처리
    FinisherGroupUpdate:
      properties:
        batch_name:
          anyOf:
          - type: string
            maxLength: 100
          - type: 'null'
          title: Batch Name
        head_count_in:
          anyOf:
          - type: integer
            minimum: 1.0
          - type: 'null'
          title: Head Count In
        avg_entry_weight_kg:
          anyOf:
          - type: number
            exclusiveMinimum: 0.0
          - type: 'null'
          title: Avg Entry Weight Kg
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
      type: object
      title: FinisherGroupUpdate
      description: 그룹 수정 (입식 정보)
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    PigletGroupCreate:
      properties:
        group_code:
          type: string
          maxLength: 30
          title: Group Code
          description: '그룹 코드 (예: PG-2026-001)'
        batch_name:
          anyOf:
          - type: string
            maxLength: 100
          - type: 'null'
          title: Batch Name
        weaning_date:
          type: string
          format: date
          title: Weaning Date
        head_count_in:
          type: integer
          minimum: 1.0
          title: Head Count In
          description: 이유 두수
        avg_entry_weight_kg:
          anyOf:
          - type: number
            exclusiveMinimum: 0.0
          - type: 'null'
          title: Avg Entry Weight Kg
          description: 평균 이유 체중 (kg)
        building_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Building Id
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
      type: object
      required:
      - group_code
      - weaning_date
      - head_count_in
      title: PigletGroupCreate
    PigletGroupDeathRecord:
      properties:
        head_count_dead:
          type: integer
          minimum: 1.0
          title: Head Count Dead
          description: 폐사 두수 추가
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
      type: object
      required:
      - head_count_dead
      title: PigletGroupDeathRecord
      description: 자돈 폐사 기록 (누적)
    PigletGroupResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        farm_id:
          type: string
          format: uuid
          title: Farm Id
        group_code:
          type: string
          title: Group Code
        batch_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Batch Name
        weaning_date:
          type: string
          format: date
          title: Weaning Date
        transfer_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Transfer Date
        transfer_type:
          anyOf:
          - type: string
          - type: 'null'
          title: Transfer Type
        head_count_in:
          type: integer
          title: Head Count In
        head_count_dead:
          type: integer
          title: Head Count Dead
        head_count_out:
          anyOf:
          - type: integer
          - type: 'null'
          title: Head Count Out
        avg_entry_weight_kg:
          anyOf:
          - type: number
          - type: 'null'
          title: Avg Entry Weight Kg
        avg_exit_weight_kg:
          anyOf:
          - type: number
          - type: 'null'
          title: Avg Exit Weight Kg
        building_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Building Id
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - farm_id
      - group_code
      - batch_name
      - weaning_date
      - transfer_date
      - transfer_type
      - head_count_in
      - head_count_dead
      - head_count_out
      - avg_entry_weight_kg
      - avg_exit_weight_kg
      - building_id
      - notes
      - created_at
      - updated_at
      title: PigletGroupResponse
    PigletGroupTransferOut:
      properties:
        transfer_date:
          type: string
          format: date
          title: Transfer Date
        transfer_type:
          type: string
          pattern: ^(FINISHER_TRANSFER|SOLD|CULLED)$
          title: Transfer Type
        head_count_out:
          type: integer
          minimum: 1.0
          title: Head Count Out
        avg_exit_weight_kg:
          anyOf:
          - type: number
            exclusiveMinimum: 0.0
          - type: 'null'
          title: Avg Exit Weight Kg
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
      type: object
      required:
      - transfer_date
      - transfer_type
      - head_count_out
      title: PigletGroupTransferOut
      description: 전출/판매 처리
    PigletTransferCreate:
      properties:
        source_sow_id:
          type: string
          format: uuid
          title: Source Sow Id
        dest_sow_id:
          type: string
          format: uuid
          title: Dest Sow Id
        transfer_date:
          type: string
          format: date
          title: Transfer Date
        piglet_count:
          type: integer
          minimum: 1.0
          title: Piglet Count
        source_farrowing_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Source Farrowing Id
        dest_farrowing_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Dest Farrowing Id
        reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Reason
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
      type: object
      required:
      - source_sow_id
      - dest_sow_id
      - transfer_date
      - piglet_count
      title: PigletTransferCreate
      description: 양자/대리모 이벤트
    PigletTransferResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        farm_id:
          type: string
          format: uuid
          title: Farm Id
        source_sow_id:
          type: string
          format: uuid
          title: Source Sow Id
        dest_sow_id:
          type: string
          format: uuid
          title: Dest Sow Id
        transfer_date:
          type: string
          format: date
          title: Transfer Date
        piglet_count:
          type: integer
          title: Piglet Count
        source_farrowing_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Source Farrowing Id
        dest_farrowing_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Dest Farrowing Id
        reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Reason
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - farm_id
      - source_sow_id
      - dest_sow_id
      - transfer_date
      - piglet_count
      - source_farrowing_id
      - dest_farrowing_id
      - reason
      - created_at
      title: PigletTransferResponse
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
servers:
- url: https://api.pigos.io
  description: Production
- url: http://localhost:8000
  description: Local dev (iOS 시뮬)
- url: http://10.0.2.2:8000
  description: Local dev (Android 에뮬)